rename benchmark.context_storage to db_benchmark.benchmark
