--- dff/utils/benchmark/__init__.py
@@ -1,3 +1,0 @@
-# -*- coding: utf-8 -*-
-# flake8: noqa: F401
-from dff.utils.benchmark.context_storage import report as context_storage_benchmark_report
