+++ dff/utils/db_benchmark/__init__.py
@@ -1,0 +1,3 @@
+# -*- coding: utf-8 -*-
+# flake8: noqa: F401
+from dff.utils.db_benchmark.benchmark import report as context_storage_benchmark_report
